requirements.txt
streamlit
pytesseract
Pillow
fpdf2
yagmail